Refactor: svg 파일 import 및 불필요한 &nbsp; 삭제 및 스타일 수정
