+++ src/components/search/KeywordItem/KeywordItem.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import * as Styled from '@/components/search/KeywordItem/KeywordItem.styled';
-import SearchIcon from '@/components/search/UI/SearchIcon';
+import { ReactComponent as SearchIcon } from '@/assets/search.svg';
 
 const KeywordItem = ({ item }) => {
   return (
     <Styled.RecentKeyword>
       <SearchIcon />
-      &nbsp; {item}
+      {item}
     </Styled.RecentKeyword>
   );
 };
