+++ src/components/search/KeywordItem/KeywordItem.styled.js
@@ -9,6 +9,7 @@ const RecentKeyword = styled.div`
   :hover {
     background-color: azure;
   }
+  column-gap: 10px;
 `;
 
 export { RecentKeyword };
