+++ src/components/search/SearchTab/SearchTab.styled.js
@@ -2,10 +2,10 @@ import styled from '@emotion/styled';
 
 const SearchTabContainer = styled.div`
   width: 100%;
+  height: 100vh;
   padding: 100px 0;
   display: flex;
   flex-direction: column;
-  justify-content: center;
   align-items: center;
   background-color: #cfebfc;
 `;
