+++ src/components/search/searchResult/SuggestionResult/SuggestionResult.styled.js
@@ -20,6 +20,7 @@ const InputKeyword = styled.div`
   :hover {
     background-color: azure;
   }
+  column-gap: 10px;
 `;
 
 const SubText = styled.div`
